Merging part of 10

See merge request COURSE-ASL2020/team040!11

+++ src/core/tests/stratified_resample_test.cpp
@@ -13,7 +13,7 @@ int main() {
             const size_t N_w = 5;
             double w[N_w] = {2,4,6,8,10}; //unnormalized weights
             double Neff = 10.0; //Just written to this variable
-            int keep[N_w] = {100,100,100,100,100}; //Just written to this variable
+            size_t keep[N_w] = {100,100,100,100,100}; //Just written to this variable
 
             when("I call stratified_resample(w, N_w, &Neff, keep) with a fixed seed") = [&] {
                 
@@ -22,7 +22,7 @@ int main() {
                 then("I get the the outputs w, Neff and keep I want") = [=] {
                     double exact_Neff = 4.090909090909;
                     double cumsum_w[N_w] = {2.0/30,6.0/30,12.0/30,20.0/30,1.0};
-                    int exact_keep[N_w] = {1,2,3,4,4};
+                    size_t exact_keep[N_w] = {1,2,3,4,4};
                     for (int i = 0; i < N_w; i++) {
                         double error_w = fabs(w[i]-cumsum_w[i]);
                         expect(that % error_w < 1e-12) << i;
